rollback unsignedBigInteger

+++ database/migrations/2024_10_25_004906_create_addresses_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('addresses', function (Blueprint $table) {
             $table->id();
-            $table->integer('receiver_id');
+            $table->unsignedBigInteger('receiver_id');
             $table->foreign('receiver_id')->references('id')->on('receivers')->onDelete('cascade');
             $table->string('street');
             $table->string('street_number');
