+++ database/migrations/2024_10_25_004906_create_phones_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('phones', function (Blueprint $table) {
             $table->id();
-            $table->integer('receiver_id');
+            $table->unsignedBigInteger('receiver_id');
             $table->foreign('receiver_id')->references('id')->on('receivers')->onDelete('cascade');
             $table->string('area_code');
             $table->string('phone_number');
